+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## Unreleased
 
+### Breaking
+* 🚚 Rename `actionsNamespace` to `actions`
+* 🚚 Rename `crudActionsNamespace` to `actionsIncludingCrud`
+
 ## [0.2.2] - 2017-07-20
 * 📦 Set redux-saga and immutability-helper as peer dependencies